Define #fillup and #auto_fillup! to ThreadPool
If a worker threads is died by unexpected error, reaper kills it
and remove it from workers list. And if all worker threads are died like
that, all of them will be done so. In that case there's no active healthy
worker thread, and Puma cannot respond to client requests.
To deal with this problem, ThreadPool should have mechanism to recover
worker threads until workers count reaches the minimum of thread count.